show: to avoid strange effects with --scale, mouseNav and deltaMouse variables must be floating point variables
git-svn-id: svn://svn.code.sf.net/p/slam6d/code/trunk@2081 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ include/show/callbacks_glut.h
@@ -25,7 +25,7 @@ void quit();
 
 void mouseMove(int x, int y);
 
-void mouseMoveDelta(int deltaMouseX, int deltaMouseY);
+void mouseMoveDelta(double deltaMouseX, double deltaMouseY);
 
 /** Reshape Function
  * TODO: have to describe it.
